- pearl shop improved: upgrades are shown in the stats bars under the characters - game over state added - minor changes

+++ GlobalProgress.h
@@ -11,7 +11,7 @@ struct Character {
     bool unlocked;
     std::string name;
     float health, speed, armor, strength;
-    int healthUpgrades, speedUpgrades, armorUpgrades, strengthUpgrades;
+    float healthUpgrades, speedUpgrades, armorUpgrades, strengthUpgrades;
     int price;
 };
